Profesor working with capacitaciones and titulos

+++ src/main/java/com/stacktrace/exam/educationserver/repository/ProfesorRepository.java
@@ -4,6 +4,6 @@ import com.stacktrace.exam.educationserver.entities.Profesor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
 
-public interface ProfesorRepository extends JpaRepository<Profesor, Integer> {
+public interface ProfesorRepository extends JpaRepository<Profesor, Long> {
     Profesor findByDni(@Param("dni") String dni);
 }
